Initial Separation
Move shared mutable datatypes into a config file

Provide helpers with minimal interactions and heavy-use an isolated environment to import from
